refactor brokers page organization
