+++ frontend/src/components/brokers/shared/BrokerCard.tsx
